+++ HISTORY.rst
@@ -5,13 +5,13 @@ History
 
 Current development branch
 
-0.2.0-devel (2016-??-??)
+0.2.0 (2017-01-01)
 ---------------------
 
 * First usable release.
 
 
-0.1.0-devel (2014-11-01)
+0.1.0 (2014-11-01)
 ---------------------
 
 * First release on PyPI.